refactor: 고차 배열 객체화, resolveCarsInfo depth 축소를 위해 함수 추출
