+++ src/main/kotlin/racingcar/domain/GameWinnerSelector.kt
@@ -1,5 +1,7 @@
 package racingcar.domain
 
+import racingcar.domain.history.CarHistory
+
 object GameWinnerSelector {
     fun decideWinners(histories: List<CarHistory>): List<String> {
         val maxLocation = histories.maxOf { it.locationValue }
