+++ src/main/kotlin/racingcar/domain/car/CarMoveStrategy.kt
@@ -1,4 +1,4 @@
-package racingcar.domain
+package racingcar.domain.car
 
 import racingcar.util.NumberGenerator
