+++ src/main/kotlin/racingcar/domain/history/CarHistories.kt
@@ -1,0 +1,4 @@
+package racingcar.domain.history
+
+@JvmInline
+value class CarHistories(val carHistories: List<CarHistory>)
