+++ src/main/kotlin/racingcar/domain/history/CarHistory.kt
@@ -1,3 +1,3 @@
-package racingcar.domain
+package racingcar.domain.history
 
 class CarHistory(val nameValue: String, val locationValue: Int)
