+++ src/main/kotlin/racingcar/view/ResultView.kt
@@ -1,23 +1,28 @@
 package racingcar.view
 
-import racingcar.domain.CarHistory
+import racingcar.domain.history.CarHistories
+import racingcar.domain.history.CarHistory
 
 class ResultView {
     fun resolveTitle() {
         println("실행 결과")
     }
 
-    fun resolveCarsInfo(data: List<List<CarHistory>>) {
-        data.forEach { carHistories ->
-            carHistories.forEach { carHistory ->
-                print("${carHistory.nameValue} : ")
-                println("-".repeat(carHistory.locationValue))
-            }
-            println()
-        }
+    fun resolveCarsInfo(data: MutableList<CarHistories>) {
+        data.forEach { printCarHistories(it) }
     }
 
     fun resolveGameWinner(data: GameWinnerResultViewModel) {
         println("${data.winners.joinToString(", ")}가 최종 우승했습니다.")
+    }
+
+    private fun printCarHistories(carHistories: CarHistories) {
+        carHistories.carHistories.forEach { printCarHistory(it) }
+        println()
+    }
+
+    private fun printCarHistory(carHistory: CarHistory) {
+        print("${carHistory.nameValue} : ")
+        println("-".repeat(carHistory.locationValue))
     }
 }
