+++ src/test/kotlin/racingcar/domain/RacingGameTest.kt
@@ -2,6 +2,7 @@ package racingcar.domain
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
+import racingcar.domain.car.Cars
 import racingcar.util.FakeRandomGenerator
 
 class RacingGameTest : StringSpec({
